Fix checkstyle errors
