--- src/main/java/Angela/task/ToDoTask.java
@@ -1,20 +1,0 @@
-package Angela.task;
-
-
-
-
-public class ToDoTask extends Task {
-	public ToDoTask(String name) {
-		super(name);
-	}
-
-	@Override
-	public String toString() {
-		return "[T]" + super.toString();
-	}
-
-	@Override
-	public String stringify() {
-		return "T||" + super.stringify();
-	}
-}
