+++ src/main/java/angelapackage/exception/InvalidDateTimeAngelaException.java
@@ -1,4 +1,4 @@
-package Angela.exception;
+package angelapackage.exception;
 
 public class InvalidDateTimeAngelaException extends AngelaException {
     public InvalidDateTimeAngelaException() {
