+++ src/main/java/angelapackage/exception/MissingArgumentAngelaException.java
@@ -1,4 +1,4 @@
-package Angela.exception;
+package angelapackage.exception;
 
 public class MissingArgumentAngelaException extends AngelaException {
