+++ src/main/java/angelapackage/exception/OutOfBoundsAngelaException.java
@@ -1,4 +1,4 @@
-package Angela.exception;
+package angelapackage.exception;
 
 public class OutOfBoundsAngelaException extends AngelaException {
